feat: naive implementation of a cursor
Problem: we can't have an actual '|' character as the cursor as this
would cause the ui to shift because we are introducing a new character
into the mix. Another solution is to have some sort of ghost text or
some sort of base layer + cursor layer.

Is getting late to think about this, this is future me problem.