fix: add flag to overwrite color support
Relates to #10. This is a temporary fix as I haven't found a clean way
to detect true color support. For example, our simple check detect the
lack of color support in macOS Terminal.app correctly. But if we enter a tmux
session with that same Terminal.app it will, incorrectly detect,
`truecolor` support as my own tmux config sets it that way (to handle
some edge cases that are not revelant here) with `set -g default-terminal "tmux-256color"`

Eventaully we might do a more robust check, but this is good for now as
it, at least, allows the user to overwrite the color support. The only
thing bothering me still is that potential bad first impression in cases
were the user is using one such terminal that do not have proper color
support and using Tmux session with flags overwritting `$COLORTERM`.