fix: correctly load langugage pack using assets.rs
Before we were loading the language pack using the filesystem.
This was wrong but I didn't noticed the problem as I was executing
the `termitype` binary from the root of the repo everytime and
everything seem to be working fine. It wasn't until I downloaded the
package from crates.io on another computer (which didn't had the repo)
that I found to cause of the missing language pack on installed
binaries.