fix: don't hide cursor when menu is opened
Before, hiding the cursor was desired as we had a floating menu on the
middle of screen and the cursor was displayed on top of the floating
menu. Now with the quake style menu on the top we don't have the need to
hide the cursor.

In fact, now we can actually see the cursor change while using `Change
Cursor Style` on the menu.