fix: bug were correct word was marked as wrong
This fixes a literal edge case where a correctly
typed word was marked as a wrongly typed word.

If had the following test words case "hey"

and if you typed the following sequence:

- 'h'
- 'e'
- <space>
- <backspace>
- 'y'

This resulted in the correctly typed word "hey"
but regardless of this, it was still being marked
as wrong after jumping to the next word.