chore: vendor the themes in the repo
Before we downloaded the themes at build time which saved
us from having all the themes at the repo.

The downside of this is that it resulted in a dirty build which
wasn't going to work with package managers like `nixpkgs`.

Maybe another solution would be to only package "essential" themes
on with the application and have the users that want more themes
to manually click `download` in the application or via a cli
flag such as `termitype --download-themes`. This is an option for the
future.

For now, do the lazy thing which is pakcage all the themes in the repo.