fix: don't close the menu when opening the leaderboards
Before we closed the `Menu` entirely after opening the `Leaderboard`.
We did this because it let the leaderboard take the spotlight without having
the menu be distracting on the background.

The downside of the above approach was that when the user closed the `Leaderboard`
it didn't show the menu again, which felt discontinuous.

Now, if the leaderboard is opened through the menu, then after closing the leaderboard we show
the menu in the state it was before opening the leaderboard.

One possible downside to this is that we just "hide" the menu while the leaderboard is opened and
having the menu opened at the same time the leaderboard is opened might be a little wasteful. I don't think
is enough to warrant having a sort of flag that determines if the leaderboard was opened via the menu
or the global key map for `LeaderboardToggle`