fix: theme menu always used the intial theme
Fixes a bug where the index resolver of the current theme
was using the wrong theme reference causing the selected
Theme menu index to always be wrong after we changed a theme.

We were using the `config.theme` and we only sync the changes to
the config after the `Quit` event, thus this caused the bug
of always opening the theme menu on the theme that was set
on launch as per the config state.